+++ src/HardenedFetch.spec.ts
@@ -35,13 +35,15 @@ describe('Hardened Fetch', () => {
 
       const instance = new HardenedFetch()
 
-      instance.fetch('http://www.example.com/')
+      instance.queue.pause()
 
-      assert.equal(instance.queue.jobs().length, 1)
+      const promise = instance.fetch('http://www.example.com/')
 
-      await new Promise((resolve) => {
-        instance.queue.on('done', resolve)
-      })
+      assert.equal(instance.queue.pending, 1)
+
+      instance.queue.resume()
+
+      await promise
     })
 
     it('resolves with the Response object', async () => {
@@ -54,10 +56,11 @@ describe('Hardened Fetch', () => {
     })
 
     it('rejects with a HTTP error on failure', async () => {
-      mockClient.intercept({ path: '/' }).reply(404)
+      mockClient.intercept({ path: '/' }).reply(404, 'Not Found')
 
       const instance = new HardenedFetch()
-      await assert.rejects(() => instance.fetch('http://www.example.com/'), /NotFound/)
+
+      await assert.rejects(() => instance.fetch('http://www.example.com/'), /HTTP error: 404/)
     })
 
     it('appends base URL when defined', async () => {
@@ -78,46 +81,6 @@ describe('Hardened Fetch', () => {
       const response = await instance.fetch('http://www.example.com/')
 
       assert.ok(response.ok)
-    })
-  })
-
-  describe('.paginatedFetch()', () => {
-    it('returns an async iterator', async () => {
-      mockClient.intercept({ path: '/1' }).reply(200, 'OK', {
-        headers: { link: '<http://www.example.com/2>; rel="next"' },
-      })
-      mockClient.intercept({ path: '/2' }).reply(200, 'OK', {
-        headers: { link: '<http://www.example.com/3>; rel="next"' },
-      })
-      mockClient.intercept({ path: '/3' }).reply(200, 'OK')
-
-      const instance = new HardenedFetch()
-
-      const pages = instance.paginatedFetch('http://www.example.com/1')
-
-      const responses: boolean[] = []
-
-      for await (const { response, done } of pages) {
-        assert.ok(response instanceof Response)
-        responses.push(done)
-      }
-
-      assert.deepEqual(responses, [false, false, true])
-    })
-
-    it('rejects with an HTTP error on failure', async () => {
-      mockClient.intercept({ path: '/1' }).reply(200, 'OK', {
-        headers: { link: '<http://www.example.com/2>; rel="next"' },
-      })
-      mockClient.intercept({ path: '/2' }).reply(404, 'Not Found')
-
-      const instance = new HardenedFetch()
-
-      const pages = instance.paginatedFetch('http://www.example.com/1')
-
-      await pages.next()
-
-      await assert.rejects(() => pages.next(), /NotFound/)
     })
   })
 })