+++ src/handleFailed.spec.ts
@@ -1,28 +1,26 @@
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
-import createHttpError from 'http-errors'
-import { handleFailed } from './handleFailed.js'
-import type { Options } from './options.js'
+import { handleFailed, type Options } from './handleFailed.js'
+import { HTTPError } from './errors.js'
 
 const options: Options = {
-  // Throttle options
-  maxConcurrency: 0,
-  minRequestTime: 100,
   // Retry options
   maxRetries: 3,
-  doNotRetry: [404],
+  doNotRetryMethods: ['POST'],
+  doNotRetryCodes: [404],
   // Rate limit options
-  rateLimitHeader: 'Retry-After',
-  resetFormat: 'seconds',
+  rateLimitHeaders: ['Retry-After'],
 }
 
-const createError = (status: number, headers = {}) => {
+const createHttpError = (status: number, headers = {}, method = 'GET') => {
+  const request = new Request('http://www.example.com', { method })
   const response = new Response(null, { status, headers })
-  return createHttpError(response.status, { response })
+
+  return new HTTPError(request, response)
 }
 
-const createTimeout = () => {
-  class TimeoutError extends Error {
+const createTimeoutError = () => {
+  class TimeoutError extends TypeError {
     constructor() {
       super('The operation was aborted due to timeout')
       this.name = 'TimeoutError'
@@ -32,24 +30,43 @@ const createTimeout = () => {
   return new TimeoutError()
 }
 
+const createNetworkError = (code: string) => {
+  class NetworkError extends TypeError {
+    public code: string
+
+    constructor(code) {
+      super('fetch failed')
+      this.name = 'NetworkError'
+      this.code = code
+    }
+  }
+
+  return new NetworkError(code)
+}
+
 describe('Handle Failed', () => {
   describe('Retries', () => {
     it('exponentially increases the wait time as retry count increases', () => {
-      const httpError = createError(500)
+      const httpError = createHttpError(500)
       assert.equal(handleFailed(options, httpError, 0), 1000)
       assert.equal(handleFailed(options, httpError, 1), 4000)
       assert.equal(handleFailed(options, httpError, 2), 9000)
 
-      const timeoutError = createTimeout()
+      const timeoutError = createTimeoutError()
       assert.equal(handleFailed(options, timeoutError, 0), 1000)
       assert.equal(handleFailed(options, timeoutError, 1), 4000)
       assert.equal(handleFailed(options, timeoutError, 2), 9000)
+
+      const networkError = createNetworkError('ECONNRESET')
+      assert.equal(handleFailed(options, networkError, 0), 1000)
+      assert.equal(handleFailed(options, networkError, 1), 4000)
+      assert.equal(handleFailed(options, networkError, 2), 9000)
     })
 
     describe('when retry count reaches the limit', () => {
       it('returns nothing', () => {
-        const httpError = createError(500)
-        const timeoutError = createTimeout()
+        const httpError = createHttpError(500)
+        const timeoutError = createTimeoutError()
 
         assert.equal(handleFailed(options, httpError, 3), undefined)
         assert.equal(handleFailed(options, timeoutError, 3), undefined)
@@ -58,7 +75,15 @@ describe('Handle Failed', () => {
 
     describe('when response status is flagged as do not retry', () => {
       it('returns nothing', () => {
-        const error = createError(404)
+        const error = createHttpError(404)
+
+        assert.equal(handleFailed(options, error, 3), undefined)
+      })
+    })
+
+    describe('when request method is flagged as do not retry', () => {
+      it('returns nothing', () => {
+        const error = createHttpError(500, {}, 'POST')
 
         assert.equal(handleFailed(options, error, 3), undefined)
       })
@@ -67,19 +92,19 @@ describe('Handle Failed', () => {
 
   describe('Rate limiting', () => {
     it('adds 1 second to the reset time', () => {
-      const error = createError(429, { 'Retry-After': '12' })
+      const error = createHttpError(429, { 'Retry-After': '12' })
       assert.equal(handleFailed(options, error, 0), 13000)
     })
 
     describe('when no reset header is found', () => {
       it('returns default retry value', () => {
-        const error = createError(429)
+        const error = createHttpError(429)
         assert.equal(handleFailed(options, error, 0), 1000)
       })
     })
   })
 
-  describe('with an unsupported error', () => {
+  describe('with any other error', () => {
     it('returns nothing', () => {
       const error = new Error()
       assert.equal(handleFailed(options, error, 0), undefined)