+++ src/index.ts
@@ -1,2 +1,3 @@
 export { HardenedFetch } from './HardenedFetch.js'
+export * from './errors.js'
 export type { Options } from './options.d.ts'