+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v2.0.0
+
+- Node v20 is now the minimum supported version
+- Removed `.paginatedFetch()` method as there are too many variants to reasonably support
+- Removed `resetFormat` option as the library will now detect date strings/timestamps/deltas in seconds or milliseconds
+- Renamed `doNotRetry` option to `doNotRetryCodes` to differentiate from new options
+- Added `doNotRetryMethods` option to skip retrying specific request methods
+- Added support for combining abort signals passed via `.fetch()` options
+- Added support for retrying some, possibly temporary, network errors (e.g. `ECONNRESET`)
+- Switched `bottleneck` dependency to `simple-rate-limited-queue` as Bottleneck has not been maintained for several years
+- Refactored options to support new `simple-rate-limited-queue` dependency
+- Refactored `rateLimitHeader` to `rateLimitHeaders` to support multiple names
+
 ## v1.2.3
 
 - Refactored URL joining to allow path on `baseUrl` option